# backend/docs/swagger.json
{
    "components": {
        "schemas": {
...
            "httpx.ErrorDetail": {
                "properties": {
                    "code": {
                        "example": "invalid_request",
                        "type": "string"
                    },
                    "message": {
                        "example": "Invalid request",
                        "type": "string"
                    }
                },
                "required": [
                    "code",
                    "message"
                ],
                "type": "object"
            },
            "httpx.ErrorResponse": {
                "properties": {
                    "error": {
                        "$ref": "#/components/schemas/httpx.ErrorDetail"
                    }
                },
                "required": [
                    "error"
                ],
                "type": "object"
            },
...
            "profile.profileResponse": {
                "properties": {
                    "avatar_url": {
                        "type": "string"
                    },
                    "created_at": {
                        "format": "date-time",
                        "type": "string"
                    },
                    "email": {
                        "format": "email",
                        "type": "string"
                    },
                    "id": {
                        "format": "uuid",
                        "type": "string"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "avatar_url",
                    "created_at",
                    "email",
                    "id",
                    "name"
                ],
                "type": "object"
            },
            "profile.updateRequest": {
                "properties": {
                    "name": {
                        "description": "The 100-character limit applies after trimming, not to the raw JSON string.",
                        "example": "Sergey",
                        "type": "string"
                    }
                },
                "required": [
                    "name"
                ],
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "bearerFormat": "JWT",
                "description": "Access JWT issued by registration, login, or refresh.",
                "scheme": "bearer",
                "type": "http"
            },
            "RefreshCookie": {
                "description": "HttpOnly refresh cookie scoped to /api/v1/auth. Sessions expire 30 days after login; rotation does not extend expiry.",
                "in": "cookie",
                "name": "refresh_token",
                "type": "apiKey"
            }
        }
    },
    "info": {
        "description": "Authentication, profiles, avatars, and website monitor configuration. Unknown JSON fields are rejected. Monitoring checks are not implemented.\nRequests with Origin must match ALLOWED_ORIGIN. Mutating requests without Origin require X-CSRF-Protection: 1. Access JWTs expire after 900 seconds; logout does not invalidate existing JWTs.",
        "title": "Uptime API",
        "version": "1.0.0"
    },
...
    "paths": {
...
        "/avatars/{filename}": {
            "get": {
                "description": "Serves a public JPEG or PNG by its generated lowercase UUID filename. Supports conditional and range requests through net/http ServeContent. Files are publicly cacheable for one year.",
                "operationId": "getAvatar",
                "parameters": [
                    {
                        "description": "Lowercase UUID followed by .png or .jpg",
                        "example": "550e8400-e29b-41d4-a716-446655440000.png",
                        "in": "path",
                        "name": "filename",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional byte range",
                        "example": "bytes=0-1023",
                        "in": "header",
                        "name": "Range",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Optional HTTP timestamp for cache validation",
                        "in": "header",
                        "name": "If-Modified-Since",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "HTTP timestamp; serve the range only if the image is unchanged",
                        "in": "header",
                        "name": "If-Range",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "image/jpeg": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "image/png": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "JPEG or PNG image bytes",
                        "headers": {
                            "Cache-Control": {
                                "description": "public, max-age=31536000, immutable",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Last-Modified": {
                                "description": "HTTP timestamp for conditional requests",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "206": {
                        "content": {
                            "image/jpeg": {
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "image/png": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        },
                        "description": "Partial image bytes",
                        "headers": {
                            "Cache-Control": {
                                "description": "public, max-age=31536000, immutable",
                                "schema": {
                                    "type": "string"
                                }
                            },
                            "Last-Modified": {
                                "description": "HTTP timestamp for conditional requests",
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "304": {
                        "description": "Image not modified"
                    },
                    "403": {
                        "description": "JSON error envelope: Origin rejected"
                    },
                    "404": {
                        "description": "Plain text: avatar not found"
                    },
                    "412": {
                        "description": "Plain text: conditional request precondition failed"
                    },
                    "416": {
                        "description": "Plain text: unsatisfiable byte range"
                    }
                },
                "summary": "Get a public avatar",
                "tags": [
                    "profile"
                ]
            }
        },
...
        "/profile": {
            "get": {
                "operationId": "getProfile",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/profile.profileResponse"
                                }
                            }
                        },
                        "description": "Current profile"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token or deleted user"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unable to load profile"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Get the profile",
                "tags": [
                    "profile"
                ]
            },
            "patch": {
                "description": "Accepts only name. Surrounding whitespace is trimmed; at most 100 Unicode characters without controls are allowed. An empty string clears the name. JSON body is limited to 4096 bytes.",
                "operationId": "updateProfile",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/profile.updateRequest",
                                "summary": "profile",
                                "description": "New name; empty string clears it"
                            }
                        }
                    },
                    "description": "New name; empty string clears it",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/profile.profileResponse"
                                }
                            }
                        },
                        "description": "Updated profile"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid JSON or name"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token or deleted user"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unable to save profile"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update the profile name",
                "tags": [
                    "profile"
                ]
            }
        },
        "/profile/avatar": {
            "post": {
                "description": "Requires exactly one name field and one avatar file. Name is trimmed and limited to 100 Unicode characters without controls; empty clears it. JPEG/PNG files are limited to 5 MiB and 2048 by 2048 pixels. Images are re-encoded without uploaded metadata.",
                "operationId": "uploadAvatar",
                "parameters": [
                    {
                        "description": "Allowed frontend origin; required for browser requests",
                        "in": "header",
                        "name": "Origin",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "description": "Must be 1 for mutating requests without Origin",
                        "in": "header",
                        "name": "X-CSRF-Protection",
                        "schema": {
                            "enum": [
                                "1"
                            ],
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "description": "JPEG or PNG, at most 5 MiB and 2048 by 2048 pixels",
                                        "format": "binary",
                                        "type": "string"
                                    },
                                    "name": {
                                        "description": "Profile name; at most 100 Unicode characters after trimming; empty string clears it",
                                        "type": "string"
                                    }
                                },
                                "required": [
                                    "name",
                                    "avatar"
                                ],
                                "type": "object"
                            }
                        }
                    },
                    "description": "Profile name; at most 100 Unicode characters after trimming; empty string clears it",
                    "required": true
                },
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/profile.profileResponse"
                                }
                            }
                        },
                        "description": "Updated profile"
                    },
                    "400": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid multipart fields, name, or image"
                    },
                    "401": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Invalid access token or deleted user"
                    },
                    "403": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Origin or CSRF rejected"
                    },
                    "413": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Avatar or multipart body too large"
                    },
                    "500": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/httpx.ErrorResponse"
                                }
                            }
                        },
                        "description": "Unable to save avatar"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "summary": "Update the name and avatar",
                "tags": [
                    "profile"
                ]
            }
        }
    },
    "openapi": "3.1.0",
    "servers": [
        {
            "description": "Current API server",
            "url": "/api/v1"
        }
    ]
}